Create README.md
Primarily here to collaborate with people for the end goal of us making good honest large sums of money without hurting the planet or swindling people... is this a good place for that? If you are old fashioned you can also reach me by inserting, in address form: arrive ~  at -  love - dot - com |  into your favorite electronic-mail (or e-mail) application's TO: field and then pressing the <return> bar or button and other fields that may be applicable depending what it is you would like me to read from you and then using your mouse to hit the SEND button (which usually has a little icon of an envelope ((an ancient origami container for a way of texting someone) sometimes garnished with little squiggled lines that are meant to indicate motion in an adjacent pictogram)).